refactored logic to use a 64 bit hash function

+++ inc/numeral_element_types.h
@@ -14,7 +14,7 @@ int can_compare_numeral_type_element_def(const element_def* ele_d_1, const eleme
 
 int compare_numeral_type_elements(const void* e1, const element_def* ele_d_1, const void* e2, const element_def* ele_d_2);
 
-uint32_t hash_numeral_type_element(const void* e, const element_def* ele_d, uint32_t (*hash_func)(const void* data, uint32_t size));
+uint64_t hash_numeral_type_element(const void* e, const element_def* ele_d, uint64_t (*hash_func)(const void* data, uint32_t size));
 
 void set_numeral_element(void* e, const element_def* ele_d, const user_value* uval);
